requirements.txt
# Core dependencies from original CoCoNuT
torch>=2.5.1
numpy>=2.1.3
transformers>=4.46.2
wandb>=0.18.7
datasets>=3.1.0
tqdm>=4.67.0

# Additional dependencies for multimodal functionality
Pillow>=10.0.0
opencv-python>=4.8.0
timm>=0.9.0

# InternVL specific dependencies
flash-attn>=2.3.6
accelerate>=0.20.0
deepspeed>=0.10.0

# Data processing and evaluation
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration and utilities
pyyaml>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Optional dependencies for specific features
# lmdeploy>=0.6.4  # Uncomment if using LMDeploy for serving
# apex  # Uncomment if using NVIDIA Apex for mixed precision